+++ test/test_formats.py
@@ -2,23 +2,18 @@ import os
 import unittest
 
 from seaborn_table.table import SeabornTable, UNICODE, BASESTRING
+from test.support import BaseTest
 
-PATH = os.path.split(os.path.abspath(__file__))[0]
 
-
-class QuoteNumbersTest(unittest.TestCase):
+class QuoteNumbersTest(BaseTest):
     def quote_number_test(self, source):
-        file_path = os.path.join(PATH, 'data', 'test_file.rst')
-        table = SeabornTable.file_to_obj(file_path=file_path)
+        table = self.get_base_table()
         table.map(lambda x: UNICODE(x))
-        result_path = os.path.join(PATH, 'test_file.%s' % source)
-        result = table.obj_to_file(result_path, quote_numbers=False)
-        expected_file = os.path.join(PATH, 'data', 'expected',
-                                     'test_file.%s' % source)
-        with open(expected_file, 'rb') as fp:
-            self.assertEqual(fp.read().decode('utf-8').replace('\r', ''),
-                             result.replace('\r', ''))
-        os.remove(result_path)
+        result_file = self.test_data_path('result', 'test_file.%s' % source)
+        table.obj_to_file(result_file, quote_numbers=False)
+        expected_file = self.test_data_path('expected', 'test_file.%s' % source)
+        self.assert_file_equal(expected_file, result_file)
+        self.remove_file(result_file)
 
     def test_quote_numbers_md(self):
         self.quote_number_test('md')
@@ -47,7 +42,7 @@ class QuoteNumbersTest(unittest.TestCase):
 
 class EvalCellsFalseTest(unittest.TestCase):
     def eval_cells_test(self, source):
-        file_path = os.path.join(PATH, 'data', 'test_file.%s' % source)
+        file_path = self.test_data_path('test_file.%s' % source)
         table = SeabornTable.file_to_obj(eval_cells=False, file_path=file_path)
         assert isinstance(table[0]['TU'], BASESTRING)
 
@@ -77,23 +72,19 @@ class EvalCellsFalseTest(unittest.TestCase):
         self.eval_cells_test('psql')
 
 
-class InsertPopRemoveColumnTest(unittest.TestCase):
-    maxDiff = None
-
+class InsertPopRemoveColumnTest(BaseTest):
     def insert_pop_remove_column(self, source):
-        file_path = os.path.join(PATH, 'data', 'test_file.rst')
-        table = SeabornTable.file_to_obj(file_path=file_path)
+        table = self.get_base_table()
         table.insert(0, 'remove_column')
         table.insert(None, 'pop_empty_columns')
         table.remove_column('remove_column')
         table.pop_empty_columns()
-        result_path = os.path.join(PATH, 'test_file.%s' % source)
-        result = table.obj_to_file(result_path)
-        expected_file = os.path.join(PATH, 'data', 'test_file.%s' % source)
-        with open(expected_file, 'rb') as fp:
-            self.assertEqual(fp.read().decode('utf-8').replace('\r', ''),
-                             result.replace('\r', ''))
-        os.remove(result_path)
+        result_file = self.test_data_path('column_results',
+                                          'test_file.%s' % source)
+        table.obj_to_file(result_file)
+        expected_file = self.test_data_path('test_file.%s' % source)
+        self.assert_file_equal(expected_file, result_file)
+        self.remove_file(result_file)
 
     def test_insert_pop_remove_column_md(self):
         self.insert_pop_remove_column('md')
@@ -118,6 +109,53 @@ class InsertPopRemoveColumnTest(unittest.TestCase):
 
     def test_insert_pop_remove_column_psql(self):
         self.insert_pop_remove_column('psql')
+
+
+class SharedColumnTest(BaseTest):
+    SHARED_LIMIT = 0
+
+    def insert_pop_remove_column(self, source):
+        tables = [SeabornTable.file_to_obj(
+            file_path=self.test_data_path('test_share_columns_%s.rst' % i))
+            for i in range(4)]
+        for table in tables:
+            table.share_column_widths(tables, self.SHARED_LIMIT)
+        basename = 'test_share_columns_%s.%s' % (self.SHARED_LIMIT, source)
+        expected_file = self.test_data_path('expected', basename)
+        result_file = self.test_data_path('shared', basename)
+        with open(result_file, 'rb') as fn:
+            for table in tables:
+                fn.write(table.obj_to_type(source)+b'\n\n')
+        self.assert_file_equal(expected_file, result_file)
+        self.remove_file(result_file)
+
+    def test_share_column_md(self):
+        self.share_column('md')
+
+    def test_share_column_csv(self):
+        self.share_column('csv')
+
+    def test_share_column_txt(self):
+        self.share_column('txt')
+
+    def test_share_column_html(self):
+        self.share_column('html')
+
+    def test_share_column_grid(self):
+        self.share_column('grid')
+
+    def test_share_column_json(self):
+        self.share_column('json')
+
+    def test_share_column_rst(self):
+        self.share_column('rst')
+
+    def test_share_column_psql(self):
+        self.share_column('psql')
+
+
+class SharedColumnLimitTest(BaseTest):
+    SHARED_LIMIT = 3
 
 
 if __name__ == '__main__':